+++ packages/components/src/navigator/navigator-provider/component.tsx
@@ -27,6 +27,10 @@ import type {
 	Screen,
 	NavigateToParentOptions,
 } from '../types';
+import { NavigatorBackButton } from '../navigator-back-button';
+import { NavigatorScreen } from '../navigator-screen';
+import { NavigatorButton } from '../navigator-button';
+import { NavigatorToParentButton } from '../navigator-to-parent-button';
 
 type MatchedPath = ReturnType< typeof patternMatch >;
 
@@ -267,35 +271,32 @@ function UnconnectedNavigatorProvider(
 }
 
 /**
- * The `NavigatorProvider` component allows rendering nested views/panels/menus
- * (via the `NavigatorScreen` component and navigate between these different
- * view (via the `NavigatorButton` and `NavigatorBackButton` components or the
- * `useNavigator` hook).
+ * The `Navigator.Root` allows rendering nested views/panels/menus
+ * (via the `Navigator.Screen` component) and navigate between these
+ * different views (via the `Navigator.Button` and `Navigator.BackButton`
+ * components or the `useNavigator` hook).
  *
  * ```jsx
  * import {
- *   __experimentalNavigatorProvider as NavigatorProvider,
- *   __experimentalNavigatorScreen as NavigatorScreen,
- *   __experimentalNavigatorButton as NavigatorButton,
- *   __experimentalNavigatorBackButton as NavigatorBackButton,
+ *   Navigator,
  * } from '@wordpress/components';
  *
  * const MyNavigation = () => (
- *   <NavigatorProvider initialPath="/">
- *     <NavigatorScreen path="/">
+ *   <Navigator.Root> initialPath="/">
+ *     <Navigator.Screen path="/">
  *       <p>This is the home screen.</p>
- *        <NavigatorButton path="/child">
+ *        <Navigator.Button path="/child">
  *          Navigate to child screen.
- *       </NavigatorButton>
- *     </NavigatorScreen>
+ *       </Navigator.Button>
+ *     </Navigator.Screen>
  *
- *     <NavigatorScreen path="/child">
+ *     <Navigator.Screen path="/child">
  *       <p>This is the child screen.</p>
- *       <NavigatorBackButton>
+ *       <Navigator.BackButton>
  *         Go back
- *       </NavigatorBackButton>
- *     </NavigatorScreen>
- *   </NavigatorProvider>
+ *       </Navigator.BackButton>
+ *     </Navigator.Screen>
+ *   </Navigator.Root>
  * );
  * ```
  */
@@ -303,5 +304,4 @@ export const NavigatorProvider = contextConnect(
 	UnconnectedNavigatorProvider,
 	'NavigatorProvider'
 );
-
 export default NavigatorProvider;